Refactor monolithic pkg/agent/server_http.go into domain-specific files
